requirements.txt
django
requests
PyJWT
cryptography
python-dotenv